fix scripts for office,device,wfh request
