+++ src/data/wfhRequest.data.ts
@@ -1,22 +1,5 @@
-import { faker } from "@faker-js/faker";
-import { getRandomContent } from "./fakerUtils";
+import { getMultiFutureDates, getRandomContent } from "./fakerUtils";
 import { WFHRequestForm } from "./requestTemplate.data";
-
-function getMultiFutureDates(count) {
-  const dates = [];
-
-  for (let i = 0; i < count; i++) {
-    const date = faker.date.future();
-    const formattedDate = date.toLocaleDateString("en-GB", {
-      day: "2-digit",
-      month: "2-digit",
-      year: "numeric",
-    });
-    dates.push(formattedDate);
-  }
-
-  return dates.join(", ");
-}
 
 export const WfhRequestData = {
   user: {
