+++ src/features/deviceRequest.feature
@@ -9,7 +9,7 @@ Feature: Device Request
       Then I should see request is "pending" with title "*global[deviceRequest2].getTitle" and state "PM Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[deviceRequest2].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[deviceRequest2].getTitle"
       Then I should see request is "approve" with title "*global[deviceRequest2].getTitle" and state "PM Reviews" on tasks page
 
     Scenario: I can reject the request success
@@ -20,14 +20,14 @@ Feature: Device Request
   Rule: As IT, I want to received a Device Request after PM approved
     Background:
       Given User create "Device Request" with "*testData[random_device_request]__global[deviceRequest3]" success
-      And "PM" approve the request "*global[deviceRequest3].getTitle" with strength points "" and weekness points "" success, current state "PM Reviews" 
+      And "PM" approve the request "*global[deviceRequest3].getTitle" success and current state "PM Reviews" 
       And I am on "TaskPage"
 
     Scenario: I should see the request with pending status on my tasks
       Then I should see request is "pending" with title "*global[deviceRequest3].getTitle" and state "IT Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[deviceRequest3].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[deviceRequest3].getTitle"
       Then I should see request is "approve" with title "*global[deviceRequest3].getTitle" and state "IT Reviews" on tasks page
 
     Scenario: I can reject the request success
