+++ src/features/officeEquipmentRequest.feature
@@ -9,7 +9,7 @@ Feature: Office Equipment Request
       Then I should see request is "pending" with title "*global[oer1].getTitle" and state "Branch Manager Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[oer1].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[oer1].getTitle"
       Then I should see request is "approve" with title "*global[oer1].getTitle" and state "Branch Manager Reviews" on tasks page
 
     Scenario: I can reject the request success
@@ -20,14 +20,14 @@ Feature: Office Equipment Request
   Rule: As it, I want to received an Office Equipment Request after gdvp approved
     Background:
       Given User create "Office Equipment Request" with "*testData[random_office_equipment_request]__global[oer2]" success
-      And "GDVPDN" approve the request "*global[oer2].getTitle" with strength points "" and weekness points "" success, current state "Branch Manager Reviews"
+      And "GDVPDN" approve the request "*global[oer2].getTitle" success and current state "Branch Manager Reviews"
       And I am on "TaskPage"
 
     Scenario: I should see the request with pending status on my tasks
       Then I should see request is "pending" with title "*global[oer2].getTitle" and state "IT Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[oer2].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[oer2].getTitle"
       Then I should see request is "approve" with title "*global[oer2].getTitle" and state "IT Reviews" on tasks page
 
     Scenario: I can reject the request success
@@ -37,8 +37,8 @@ Feature: Office Equipment Request
   Rule: As user, I want to see an Office Equipment Request after it approved
     Background:
       Given User create "Office Equipment Request" with "*testData[random_office_equipment_request]__global[oer3]" success
-      And "GDVPDN" approve the request "*global[oer3].getTitle" with strength points "" and weekness points "" success, current state "Branch Manager Reviews"
-      And "it" approve the request "*global[oer3].getTitle" with strength points "" and weekness points "" success, current state "IT Reviews"
+      And "GDVPDN" approve the request "*global[oer3].getTitle" success and current state "Branch Manager Reviews"
+      And "it" approve the request "*global[oer3].getTitle" success and current state "IT Reviews"
 
     Scenario: I should see the request with approved status on my requests
       When I am on "MyRequestPage"
