+++ src/features/wfhRequest.feature
@@ -9,7 +9,7 @@ Feature: WFH Request
       Then I should see request is "pending" with title "*global[wfhRequest2].getTitle" and state "PM Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[wfhRequest2].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[wfhRequest2].getTitle"
       Then I should see request is "approve" with title "*global[wfhRequest2].getTitle" and state "PM Reviews" on tasks page
 
     Scenario: I can reject the request success
@@ -20,14 +20,14 @@ Feature: WFH Request
   Rule: As GDVP, I want to received a WFH Request after PM approved
     Background:
       Given User create "WFH Request" with "*testData[random_wfh_request]__global[wfhRequest3]" success
-      And "PM" approve the request "*global[wfhRequest3].getTitle" with strength points "" and weekness points "" success, current state "PM Reviews"
+      And "PM" approve the request "*global[wfhRequest3].getTitle" success and current state "PM Reviews"
       And I am on "TaskPage"
 
     Scenario: I should see the request with pending status on my tasks
       Then I should see request is "pending" with title "*global[wfhRequest3].getTitle" and state "Branch Manager Reviews" on tasks page
 
     Scenario: I can approve the request success
-      When I approve request by title "*global[wfhRequest3].getTitle" with strength points "" and weekness points ""
+      When I approve request by title "*global[wfhRequest3].getTitle"
       Then I should see request is "approve" with title "*global[wfhRequest3].getTitle" and state "Branch Manager Reviews" on tasks page
 
     Scenario: I can reject the request success
@@ -38,8 +38,8 @@ Feature: WFH Request
   Rule: As user, I want to see a WFH after PM, GDVP approved
     Background:
       Given User create "Change Office Request" with "*testData[random_change_office_request]__global[wfhRequest4]" success
-      And "PM" approve the request "*global[wfhRequest4].getTitle" with strength points "" and weekness points "" success, current state "PM Reviews"
-      And "GDVPDN" approve the request "*global[wfhRequest4].getTitle" with strength points "" and weekness points "" success, current state "Branch Manager Reviews"
+      And "PM" approve the request "*global[wfhRequest4].getTitle" success and current state "PM Reviews"
+      And "GDVPDN" approve the request "*global[wfhRequest4].getTitle" success and current state "Branch Manager Reviews"
 
     Scenario: I should see the request with approved status on my requests
       When I am on "MyRequestPage"
